@@ -1,51 +1,64 @@
 {
-  "proxy": "http://localhost:8080",
   "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.3",
-    "@mui/material": "^5.14.3",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "axios": "^1.6.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^5.17.1",
+    "@mui/material": "^5.17.1",
+    "@mui/x-charts": "^8.5.2",
+    "@mui/x-date-pickers": "^7.28.3",
+    "@tiptap/extension-character-count": "^2.11.7",
+    "@tiptap/extension-color": "^2.11.7",
+    "@tiptap/extension-placeholder": "^2.11.7",
+    "@tiptap/extension-text-align": "^2.11.7",
+    "@tiptap/pm": "^2.11.7",
+    "@tiptap/react": "^2.11.7",
+    "@tiptap/starter-kit": "^2.11.7",
+    "axios": "^1.7.3",
     "bad-words": "^3.0.4",
     "crypto-js": "^4.2.0",
-    "date-fns": "^2.30.0",
+    "date-fns": "^3.6.0",
+    "dayjs": "^1.11.13",
+    "dompurify": "^3.2.4",
     "font-awesome": "^4.7.0",
-    "html-react-parser": "^5.0.11",
-    "install": "^0.13.0",
+    "framer-motion": "^12.10.0",
+    "frontend": "file:",
+    "html-react-parser": "^5.1.12",
+    "iso-3166-1": "^2.1.1",
     "js-cookie": "^3.0.5",
-    "jwt-decode": "^3.1.2",
-    "lodash": "^4.17.21",
-    "npm": "^9.7.1",
+    "jwt-decode": "^4.0.0",
+    "lucide-react": "^0.484.0",
+    "npm": "^10.8.2",
     "prop-types": "^15.8.1",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.8.0",
-    "react-countup": "^6.4.2",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-avatar-editor": "^13.0.2",
+    "react-bootstrap": "^2.10.4",
+    "react-countup": "^6.5.3",
+    "react-dom": "^18.3.1",
     "react-google-recaptcha": "^3.1.0",
-    "react-motion": "^0.5.2",
-    "react-placeholder-typing": "^1.0.11",
-    "react-player": "^2.12.0",
-    "react-router-dom": "^6.14.2",
+    "react-router-dom": "^6.26.0",
     "react-scripts": "^5.0.1",
-    "react-share": "^5.0.3",
-    "react-text-collapse": "^0.6.1",
+    "react-share": "^5.1.0",
+    "react-slick": "^0.30.2",
+    "react-timeago": "^7.2.0",
+    "react-type-animation": "^3.2.0",
     "reactjs-popup": "^2.0.6",
-    "sha256": "^0.2.0",
+    "serve": "^14.2.5",
+    "slick-carousel": "^1.8.1",
+    "source-map-explorer": "^2.5.3",
+    "styled-components": "^6.1.12",
     "typeface-ibm-plex-sans": "^1.1.13",
-    "web-vitals": "^2.1.4",
-    "yarn": "^1.22.19"
+    "web-vitals": "^4.2.2"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
+ "scripts": {
+  "start": "serve -s build -l $PORT",
+  "dev": "craco start",
+  "build": "craco build",
+  "test": "craco test",
+  "eject": "react-scripts eject"
+},
   "browserslist": {
     "production": [
       ">0.2%",
@@ -59,12 +72,10 @@
     ]
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.21.5",
-    "@babel/preset-react": "^7.18.6",
-    "@iconify/react": "^4.1.0",
-    "@types/react": "^18.0.28",
-    "babel-jest": "^29.5.0",
-    "jest": "^29.5.0",
-    "react-test-renderer": "^18.2.0"
+    "@craco/craco": "^7.1.0",
+    "@types/react": "^18.3.3",
+    "eslint-plugin-import": "^2.31.0",
+    "webpack-bundle-analyzer": "^5.1.0",
+    "webpack-cli": "^6.0.1"
   }
 }
