@@ -5,8 +5,10 @@ import Filter from "bad-words";
 import LoginStyles from "../Login/login.module.css";
 import MovieButton from "../Other/btn//MovieButton/Button.js";
 import CookieManager from "../../security/CookieManager";
-import { resendAuthEmail, Message, togglePasswordVisibility, ProfanityLogic } from "../../security/Shared.js";
+import { resendAuthEmail, Message, togglePasswordVisibility, ProfanityLogic } from "../Shared/Shared.js";
 const LOGIN_ENDPOINT = process.env.REACT_APP_LOGIN_ENDPOINT;
+const API_URL = process.env.REACT_APP_BACKEND_API_URL
+
 export default function LoginLogic() {
 
   const [passwordVisible, setPasswordVisible] = useState(false);
@@ -35,7 +37,7 @@ export default function LoginLogic() {
     if (!ProfanityLogic(hasProfanity, setError)) {
 
       const response = await fetch(
-        LOGIN_ENDPOINT,
+        `${API_URL}${LOGIN_ENDPOINT}`,
         {
           method: "POST",
           headers: {
