@@ -1,13 +1,25 @@
 import React from "react";
-import PopflixLogo from "./POPFLIX_LOGO_OFFICIAL.png";
-import AnimatedLogo from "./logo.module.css";
+import CritixLogo from "./CRITIX_LOGO_OFFICIAL.webp";
+import LogoStyle from "./logo.module.css";
 import { Link } from "react-router-dom";
-export default function Logo() {
+import PropTypes from "prop-types"
+export default function Logo({ placement }) {
   return (
-    <Link to="/home">
-      <div id="logo" className={AnimatedLogo.Logo}>
-        <img src={PopflixLogo} className={AnimatedLogo.Logo} alt="logo" />
-      </div>
-    </Link>
+    <>
+      {placement === 'homepage' && (
+        <img src={CritixLogo} className={LogoStyle.HomepageLogo} alt="logo" />
+      )}
+      {placement === 'login' && (
+        <img src={CritixLogo} className={LogoStyle.LoginLogo} alt="logo" />
+      )}
+      {placement === 'navbar' && (
+        <Link to='/home'>
+          <img src={CritixLogo} className={LogoStyle.NavbarLogo} alt="logo" />
+        </Link>
+      )}
+    </>
   );
 }
+Logo.propTypes = {
+  placement: PropTypes.string
+};
