@@ -4,6 +4,9 @@ import Button from '@mui/material/Button';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
 import MenuIcon from '@mui/icons-material/Menu';
+import PremiumIcon from "../../Other/crown.webp"
+import NavStyle from './Navigation.module.css'
+
 export default function Navigation() {
   const [anchorEl, setAnchorEl] = useState(null);
   const open = Boolean(anchorEl);
@@ -14,7 +17,7 @@ export default function Navigation() {
     setAnchorEl(null);
   };
   return (
-    <>
+    <div className={NavStyle.Navigation}>
       <Button
         id="basic-button"
         aria-controls={open ? 'basic-menu' : undefined}
@@ -48,21 +51,29 @@ export default function Navigation() {
           },
         }}
       >
-        <MenuItem onClick={handleClose}><Link to="/now_playing/">
-          IN THEATRES
-        </Link></MenuItem>
-        <MenuItem onClick={handleClose}>  <Link to="/upcoming/">
-          UPCOMING
-        </Link></MenuItem>
-        <MenuItem onClick={handleClose}> <Link to="/popular/">
-          MOST POPULAR
+        <Link to="/now_playing/" onClick={handleClose} style={{ textDecoration: "none" }}>
+          <MenuItem>IN THEATRES</MenuItem>
         </Link>
-        </MenuItem>
-        <MenuItem onClick={handleClose}> <Link to="/watchlist">
-          WATCHLIST
-        </Link> </MenuItem>
+        <Link to="/upcoming/" onClick={handleClose} style={{ textDecoration: "none" }}>
+          <MenuItem>UPCOMING</MenuItem>
+        </Link>
+        <Link to="/popular/" onClick={handleClose} style={{ textDecoration: "none" }}>
+          <MenuItem>MOST POPULAR</MenuItem>
+        </Link>
+        <Link to="/watchlist" onClick={handleClose} style={{ textDecoration: "none" }}>
+          <MenuItem>WATCHLIST</MenuItem>
+        </Link>
+        <MenuItem onClick={handleClose}> <Link to="/recommendations">
+          <img src={PremiumIcon} alt="Premium" style={{ width: '20px', marginRight: '8px' }} /> RECOMMENDED
+        </Link></MenuItem>
+        <Link to="/discover-movies" onClick={handleClose} style={{ textDecoration: "none" }}>
+          <MenuItem>
+            <img src={PremiumIcon} alt="Premium" style={{ width: '20px', marginRight: '8px' }} />
+            DISCOVER
+          </MenuItem>
+        </Link>
       </Menu>
-    </>
+    </div>
   );
 }
 
