@@ -1,12 +1,14 @@
 import React, { useState, useEffect, useRef } from "react";
 import SearchStyle from "./Search.module.css";
 import PropTypes from "prop-types";
-import { ParseYear } from "../../IndMovie/MovieComponents";
-import ReactPlaceholderTyping from 'react-placeholder-typing'
+import { ParseYear } from "../../IndFilm/MovieComponents";
+import { TypeAnimation } from 'react-type-animation';
 import isTokenExpired from "../../../security/IsTokenExpired";
 import { fetchData } from "../../../security/Data";
 const searchEndpoint = process.env.REACT_APP_SEARCH_ENDPOINT;
 const miniPosterUrl = process.env.REACT_APP_MINI_POSTER_URL;
+const API_URL = process.env.REACT_APP_BACKEND_API_URL
+const indMovieEndpoint = process.env.REACT_APP_IND_MOVIE_ENDPOINT
 
 export default function Search(props) {
   const [query, setQuery] = useState("");
@@ -16,7 +18,7 @@ export default function Search(props) {
   const placeholders = [
     'Discover cinematic brilliance',
     'Unearth hidden gems',
-    'Share your filmic insights',
+    'Share your film critiques',
     'Find your next favorite movie',
     'Explore cinematic wonders',
     'Express your movie passion',
@@ -53,10 +55,10 @@ export default function Search(props) {
     try {
       await isTokenExpired();
       const formattedQuery = searchQuery.includes(' ') ? searchQuery.trim().split(' ').join('+') : searchQuery.trim();
-      const endpoint = `${searchEndpoint}/${formattedQuery}`;
+      const endpoint = `${API_URL}${searchEndpoint}${formattedQuery}`;
 
       const search = await fetchData(endpoint);
-      setMovieResults(search);
+      setMovieResults(search || []);
     } catch (error) {
       console.error("Error fetching data:", error);
     }
@@ -64,20 +66,45 @@ export default function Search(props) {
 
   return (
     <form onSubmit={props.onSubmit} className={SearchStyle.Search} ref={searchRef}>
-      <ReactPlaceholderTyping
-        placeholders={placeholders}
-        value={query}
-        onChange={(value) => {
-          setQuery(value)
-        }}
-        containerStyle={{ borderWidth: "0px", justifyContent: "center" }}
-      />
+      <div style={{ position: "relative", display: "inline-block" }}>
+        {query === "" && (
+          <TypeAnimation
+            sequence={placeholders}
+            speed={20}
+            style={{
+              position: "absolute",
+              color: "gray",
+              pointerEvents: "none",
+              left: "10px",
+              top: "50%",
+              transform: "translateY(-50%)",
+            }}
+            repeat={Infinity}
+          />
+        )}
+
+        {/* Search Input */}
+        <input
+          type="text"
+          value={query}
+          onChange={(e) => setQuery(e.target.value)}
+          placeholder=""
+          style={{
+            width: "300px",
+            padding: "10px",
+            fontSize: "1em",
+            border: "1px solid #ccc",
+            borderRadius: "5px",
+            outline: "none",
+          }}
+        />
+      </div>
       <div className={SearchStyle["search-results-list"]}>
         {movieResults.map((movie) => {
           if (movie.poster_path && movie.vote_average) {
             return (
               <a
-                href={`/movies/movie/${movie.id}`}
+                href={`${indMovieEndpoint}${movie.id}`}
                 key={movie.id}
                 onClick={() => setMovieResults([])}
               >
