@@ -1,63 +1,55 @@
 import React from "react"
-import AddIcon from '@mui/icons-material/Add';
-import CheckIcon from '@mui/icons-material/Check';
 import WatchListBtnStyle from "./WatchListBtn.module.css"
+import BookmarkBorderIcon from '@mui/icons-material/BookmarkBorder';
+import BookmarkIcon from '@mui/icons-material/Bookmark';
 import { useState } from "react";
 import { sendData } from "../../../../security/Data";
 import PropTypes from "prop-types";
-export default function WatchListBtn({ movieData, userId, sendToWatchListEndpoint, deleteFromWatchListEndpoint }) {
-    const [isAddedToWatchlist, setIsAddedToWatchlist] = useState(movieData.isSavedToWatchlist);
+const sendToWatchListEndpoint = process.env.REACT_APP_ADD_TO_WATCHLIST_ENDPOINT;
+const deleteFromWatchListEndpoint = process.env.REACT_APP_DELETE_FROM_WATCHLIST_ENDPOINT;
+const API_URL = process.env.REACT_APP_BACKEND_API_URL
 
-    const handleClick = async () => {
-        if (!isAddedToWatchlist) {
-            const data = {
-                movieId: movieData.id,
-                title: movieData.title,
-                posterUrl: movieData.posterUrl,
-                voteAverage: movieData.voteAverage,
-                genres: movieData.genres,
-                overview: movieData.overview,
-                actors: movieData.actors ? movieData.actors.slice(0, 3) : null,
-                userId: userId
-            };
+export default function WatchListBtn({ movieData, outline }) {
+    const [isSavedToWatchList, setIsSavedToWatchList] = useState(movieData.isSavedToWatchlist)
 
-            const response = await sendData(`${sendToWatchListEndpoint}${userId}`, data);
-            response.ok ? setIsAddedToWatchlist(true) : setIsAddedToWatchlist(false);
-        } else {
-            const response = await sendData(`${deleteFromWatchListEndpoint}${userId}/${movieData.id}`);
-            response.ok ? setIsAddedToWatchlist(false) : setIsAddedToWatchlist(true);
+    async function handleSaveToWatchlist(e) {
+        e.preventDefault();
+        const response = await sendData(`${API_URL}${sendToWatchListEndpoint}`, movieData);
+        response.ok ? setIsSavedToWatchList(true) : false;
+        e.stopPropagation();
+    }
 
-        }
-    };
+    async function handleDeleteFromWatchlist(e) {
+        e.preventDefault();
+        const response = await sendData(`${API_URL}${deleteFromWatchListEndpoint}${movieData.movieId}`);
+        response.ok ? setIsSavedToWatchList(false) : false;
+        e.stopPropagation();
+    }
 
     return (
-        <div className={WatchListBtnStyle["btn-container"]}>
-            <div className={WatchListBtnStyle['btn']} tabIndex="0" onClick={handleClick}>
-                <div className={WatchListBtnStyle['add-icon']}>
-                    {isAddedToWatchlist ?
-                        <CheckIcon />
-                        :
-                        <AddIcon />}
-                </div>
-                <div className={WatchListBtnStyle['watchlist-label']}>
-                    <span>{isAddedToWatchlist ? 'ADDED TO WATCHLIST' : 'ADD TO WATCHLIST'}</span>
-                </div>
+        <div className={outline ? WatchListBtnStyle["circle-container"] : null}>
+            <div className={WatchListBtnStyle["action-btn"]}>
+                {!isSavedToWatchList ?
+                    <i> <BookmarkBorderIcon sx={{ fontSize: 30 }} onClick={handleSaveToWatchlist} /></i>
+                    :
+                    <i><BookmarkIcon sx={{ fontSize: 30 }} onClick={handleDeleteFromWatchlist} /></i>
+                }
             </div>
         </div>
     )
 }
+
 WatchListBtn.propTypes = {
-    userId: PropTypes.string.isRequired,
-    sendToWatchListEndpoint: PropTypes.string.isRequired,
-    deleteFromWatchListEndpoint: PropTypes.string.isRequired,
+    isSavedToWatchlist: PropTypes.bool,
+    outline: PropTypes.bool,
     movieData: PropTypes.shape({
-        id: PropTypes.number.isRequired,
-        title: PropTypes.string.isRequired,
-        posterUrl: PropTypes.string.isRequired,
-        voteAverage: PropTypes.number.isRequired,
-        genres: PropTypes.array.isRequired,
-        overview: PropTypes.string.isRequired,
+        movieId: PropTypes.number,
+        title: PropTypes.string,
+        posterUrl: PropTypes.string,
+        voteAverage: PropTypes.number,
+        genres: PropTypes.array,
+        overview: PropTypes.string,
         actors: PropTypes.array,
-        isSavedToWatchlist: PropTypes.bool.isRequired,
-    }).isRequired,
+        isSavedToWatchlist: PropTypes.bool,
+    }),
 }
