@@ -1,9 +1,13 @@
 import React from "react";
-import PopflixTrailer1 from "./PopflixTrailer1.mp4";
 import "../Login/login.module.css";
 
 export default function SignUpPlayer() {
   return (
-    <video autoPlay loop muted src={PopflixTrailer1} type="video/mp4" />
-  );
+    <iframe
+      src="https://player.vimeo.com/video/1070143041?h=70f6c51ed4&autoplay=1&loop=1&muted=1&background=1"
+      width="100%"
+      height="100%"
+      allow="autoplay; fullscreen"
+    ></iframe>
+  )
 }
