@@ -4,9 +4,9 @@ import Filter from "bad-words";
 import SignUpStyles from "../Login/login.module.css";
 import MovieButton from "../Other/btn/MovieButton/Button.js";
 import CookieManager from "../../security/CookieManager";
-import { togglePasswordVisibility, resendAuthEmail, Message, ProfanityLogic } from "../../security/Shared";
+import { togglePasswordVisibility, resendAuthEmail, Message, ProfanityLogic } from "../Shared/Shared.js";
 const SIGNUP_URL = process.env.REACT_APP_SIGNUP_ENDPOINT;
-
+const API_URL = process.env.REACT_APP_BACKEND_API_URL
 export default function SignUpFunctionality() {
 
   const [emailErr, setEmailErr] = useState(false)
@@ -19,10 +19,10 @@ export default function SignUpFunctionality() {
   const [profanityError, setProfanityError] = useState("");
   const [passwordVisible, setPasswordVisible] = useState(false);
   const [endpointResponse, setEndpointResponse] = useState(null)
-  const handleEmailChange = (event) => setEmail(event.target.value)
-  const handleFirstNameChange = (event) => setFirstName(event.target.value)
-  const handleLastNameChange = (event) => setLastName(event.target.value)
-  const handlePasswordChange = (event) => setPassword(event.target.value)
+  const handleEmailChange = (e) => setEmail(e.target.value)
+  const handleFirstNameChange = (e) => setFirstName(e.target.value)
+  const handleLastNameChange = (e) => setLastName(e.target.value)
+  const handlePasswordChange = (e) => setPassword(e.target.value)
   const handleTogglePasswordVisibility = () => {
     togglePasswordVisibility(passwordVisible, setPasswordVisible);
   };
@@ -38,47 +38,43 @@ export default function SignUpFunctionality() {
     resendAuthEmail(email, setMessage, setEmailErr);
   };
 
-  const handleSubmit = async (event) => {
-    event.preventDefault();
+  const handleSubmit = async (e) => {
+    e.preventDefault();
 
     const userData = { firstName, lastName, email, password };
     const hasProfanity = filter.isProfane(userData["firstName"]) || filter.isProfane(userData["lastName"]) || filter.isProfane(userData["email"]) || filter.isProfane(userData["password"]);
 
-    if (ProfanityLogic(hasProfanity, setProfanityError)) {
-      // Stops creation of user
-      return
-    }
+    if (!ProfanityLogic(hasProfanity, setProfanityError)) {
 
-
-    const response = await fetch(SIGNUP_URL, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(userData),
-    });
-
-    setEndpointResponse(response);
-
-    if (response.ok) {
-      const data = await response.json();
-      CookieManager.encryptCookie("accessToken", data.access_token, {
-        expires: 0.5,
+      const response = await fetch(`${API_URL}${SIGNUP_URL}`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(userData),
       });
-      CookieManager.encryptCookie("refreshToken", data.refresh_token, {
-        expires: 7,
-      });
-      setMessage(data.message);
-      resetInputFields()
-    } else {
-      const messageText = await response.text();
-      if (messageText === "There was an error sending your account activation email.") {
-        setEmailErr(true);
+
+      setEndpointResponse(response);
+
+      if (response.ok) {
+        const data = await response.json();
+        CookieManager.encryptCookie("accessToken", data.access_token, {
+          expires: 0.5,
+        });
+        CookieManager.encryptCookie("refreshToken", data.refresh_token, {
+          expires: 7,
+        });
+        setMessage(data.message);
+        resetInputFields()
+      } else {
+        const messageText = await response.text();
+        if (messageText === "There was an error sending your account activation email.") {
+          setEmailErr(true);
+        }
+        setMessage(messageText);
       }
-      setMessage(messageText);
     }
   }
-
   return (
     <form onSubmit={handleSubmit}>
       <Message response={endpointResponse} message={message} style={SignUpStyles} profanityError={profanityError} />
@@ -135,7 +131,7 @@ export default function SignUpFunctionality() {
           id="password"
           name="password"
           className={SignUpStyles["text-input"]}
-          pattern="(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{7,}"
+          pattern="(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{7,20}"
           autoComplete="current-password"
           value={password}
           onChange={handlePasswordChange}
