@@ -3,12 +3,14 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-
+import AppTheme from "./components/MUIHomepage/shared-theme/AppTheme";
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
 root.render(
   <React.StrictMode>
-    <App />
+    <AppTheme>
+      <App />
+    </AppTheme>
   </React.StrictMode>
 );
 
